requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
black==26.1.0
certifi==2026.1.4
cfgv==3.5.0
click==8.3.1
colorama==0.4.6
distlib==0.4.0
ecdsa==0.19.1
fastapi==0.128.0
filelock==3.20.3
flake8==7.3.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2==2.9.11
pyasn1==0.6.2
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.4.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
virtualenv==20.36.1
